requirements.txt
click==8.1.3
mysql_connector_repackaged==0.3.1
python-dotenv==0.20.0
setuptools==63.1.0
